Added architecture instead of just porgram.cs
