+++ FrietApp/Scraper.Logic/Abstractions/IScrapeProcess.cs
@@ -1,0 +1,6 @@
+namespace SnackApp.Logic.Scraper;
+
+public interface IScrapeProcess
+{
+    List<Models.Item> StartScrape(string baseUrl);
+}
