+++ FrietApp/Scraper.Models/Currency.cs
@@ -2,16 +2,11 @@ using System.Text.Json.Serialization;
 
 namespace SnackApp.Models;
 
-public struct Currency
+public record Currency
 {
     [JsonPropertyName("currency-symbol")]
     public string CurrencySymbol;
     
     [JsonPropertyName("value")]
     public double Value;
-
-    public override string ToString()
-    {
-        return CurrencySymbol+Value;
-    }
 }
