+++ FrietApp/Scraper.Models/ItemProperties/ItemAssociation.cs
