+++ FrietApp/Scraper.Models/ItemProperties/ItemDimension.cs
