+++ FrietApp/Scraper.Models/ItemProperties/ItemOption.cs
