+++ FrietApp/Scraper.Models/Scraper.Models.csproj
@@ -4,6 +4,7 @@
         <TargetFramework>net7.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
+        <RootNamespace>SnackApp.Models</RootNamespace>
     </PropertyGroup>
 
 </Project>
