+++ FrietApp/Scraper/Dockerfile
