--- FrietApp/SnackApp.Logic/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace SnackApp.Logic;
-
-public class Class1
-{
-}
