+++ FrietApp/SnackApp.sln
@@ -1,10 +1,12 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SnackApp", "SnackApp\SnackApp.csproj", "{0631637A-29A0-47CB-A539-D3F4E5430623}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Scraper", "Scraper\Scraper.csproj", "{0631637A-29A0-47CB-A539-D3F4E5430623}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SnackApp.Logic", "SnackApp.Logic\SnackApp.Logic.csproj", "{80504D81-367C-454C-BEB7-7C0DD9EAD5F6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Scraper.Logic", "Scraper.Logic\Scraper.Logic.csproj", "{80504D81-367C-454C-BEB7-7C0DD9EAD5F6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SnackApp.Models", "SnackApp.Models\SnackApp.Models.csproj", "{DDA2EAC2-B503-46DC-890C-654E3EFB26F4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Scraper.Models", "Scraper.Models\Scraper.Models.csproj", "{DDA2EAC2-B503-46DC-890C-654E3EFB26F4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Scraper", "Scraper", "{53D8EB85-2B2A-49D2-BB84-5F73F8F82329}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -24,5 +26,10 @@ Global
 		{DDA2EAC2-B503-46DC-890C-654E3EFB26F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DDA2EAC2-B503-46DC-890C-654E3EFB26F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DDA2EAC2-B503-46DC-890C-654E3EFB26F4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{0631637A-29A0-47CB-A539-D3F4E5430623} = {53D8EB85-2B2A-49D2-BB84-5F73F8F82329}
+		{80504D81-367C-454C-BEB7-7C0DD9EAD5F6} = {53D8EB85-2B2A-49D2-BB84-5F73F8F82329}
+		{DDA2EAC2-B503-46DC-890C-654E3EFB26F4} = {53D8EB85-2B2A-49D2-BB84-5F73F8F82329}
 	EndGlobalSection
 EndGlobal
